[minishell] UPDATE: 구조체 cmd.h로 이동, 필요없는 변수 삭제
